+++ lib/cldr/messages/exception.ex
@@ -9,3 +9,17 @@ defmodule Cldr.Message.ParseError do
     %__MODULE__{message: message}
   end
 end
+
+defmodule Cldr.Message.PositionalArgsNotPermitted do
+  @moduledoc """
+  Exception raised when parsing an ICU message format
+  and positional arguments are not permitted
+
+  """
+  defexception [:message]
+
+  def exception(message) do
+    %__MODULE__{message: message}
+  end
+end
+